app/game_service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
dj-database-url
django-environ
whitenoise
requests
django-cors-headers